+++ src/main/java/com/gdschongik/gdsc/domain/study/dto/response/StudyCurriculumResponse.java
@@ -1,6 +1,6 @@
 package com.gdschongik.gdsc.domain.study.dto.response;
 
-import com.gdschongik.gdsc.domain.recruitment.domain.vo.Period;
+import com.gdschongik.gdsc.domain.common.vo.Period;
 import com.gdschongik.gdsc.domain.study.domain.Difficulty;
 import com.gdschongik.gdsc.domain.study.domain.StudyDetail;
 import com.gdschongik.gdsc.domain.study.domain.StudyStatus;
