+++ src/main/java/com/gdschongik/gdsc/domain/study/factory/StudyDomainFactory.java
@@ -1,7 +1,7 @@
 package com.gdschongik.gdsc.domain.study.factory;
 
+import com.gdschongik.gdsc.domain.common.vo.Period;
 import com.gdschongik.gdsc.domain.member.domain.Member;
-import com.gdschongik.gdsc.domain.recruitment.domain.vo.Period;
 import com.gdschongik.gdsc.domain.study.domain.Study;
 import com.gdschongik.gdsc.domain.study.domain.StudyDetail;
 import com.gdschongik.gdsc.domain.study.dto.request.StudyCreateRequest;
