+++ src/test/java/com/gdschongik/gdsc/domain/study/application/MentorStudyAchievementServiceTest.java
@@ -3,8 +3,8 @@ package com.gdschongik.gdsc.domain.study.application;
 import static com.gdschongik.gdsc.domain.study.domain.AchievementType.*;
 import static org.assertj.core.api.Assertions.*;
 
+import com.gdschongik.gdsc.domain.common.vo.Period;
 import com.gdschongik.gdsc.domain.member.domain.Member;
-import com.gdschongik.gdsc.domain.recruitment.domain.vo.Period;
 import com.gdschongik.gdsc.domain.study.domain.Study;
 import com.gdschongik.gdsc.domain.study.domain.StudyAchievement;
 import com.gdschongik.gdsc.domain.study.dto.request.OutstandingStudentRequest;
