+++ src/test/java/com/gdschongik/gdsc/domain/study/domain/AttendanceValidatorTest.java
@@ -4,8 +4,8 @@ import static com.gdschongik.gdsc.global.common.constant.StudyConstant.ATTENDANC
 import static com.gdschongik.gdsc.global.exception.ErrorCode.*;
 import static org.assertj.core.api.Assertions.*;
 
+import com.gdschongik.gdsc.domain.common.vo.Period;
 import com.gdschongik.gdsc.domain.member.domain.Member;
-import com.gdschongik.gdsc.domain.recruitment.domain.vo.Period;
 import com.gdschongik.gdsc.global.exception.CustomException;
 import com.gdschongik.gdsc.helper.FixtureHelper;
 import java.time.LocalDate;
