+++ src/test/java/com/gdschongik/gdsc/domain/study/domain/StudyHistoryTest.java
@@ -3,8 +3,8 @@ package com.gdschongik.gdsc.domain.study.domain;
 import static com.gdschongik.gdsc.domain.study.domain.StudyHistoryStatus.*;
 import static org.assertj.core.api.Assertions.*;
 
+import com.gdschongik.gdsc.domain.common.vo.Period;
 import com.gdschongik.gdsc.domain.member.domain.Member;
-import com.gdschongik.gdsc.domain.recruitment.domain.vo.Period;
 import com.gdschongik.gdsc.helper.FixtureHelper;
 import java.time.LocalDateTime;
 import org.junit.jupiter.api.Nested;
